+++ models/user.js
@@ -34,6 +34,9 @@ const userSchema = mongoose.Schema({
   image: {
     type: String,
   },
+  secret: {
+    type: {},
+  },
   token: {
     type: String,
   },
@@ -67,7 +70,6 @@ userSchema.methods.comparepassword = function (password, cb) {
 };
 
 // generate token
-
 userSchema.methods.generateToken = function (cb) {
   var user = this;
   var token = jwt.sign(user._id.toHexString(), confiq.SECRET);
@@ -92,7 +94,6 @@ userSchema.statics.findByToken = function (token, cb) {
 };
 
 //delete token
-
 userSchema.methods.deleteToken = function (token, cb) {
   var user = this;
