+++ package.json
@@ -4,8 +4,7 @@
   "description": "medpiper backend task",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "nodemon index.js"
   },
   "author": "",
   "license": "ISC",
@@ -19,7 +18,6 @@
     "multer": "^1.4.3",
     "nodemon": "^2.0.4",
     "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
-    "path": "^0.12.7"
+    "passport-jwt": "^4.0.0"
   }
 }
