+++ routes/board/getAllBoard.js
@@ -1,0 +1,13 @@
+const board = require("../../models/board");
+module.exports = (req, res) => {
+  board.find({ userID: req.user._id }, function (err, result) {
+    if (err) {
+      console.log(err);
+      return res.status(400).json({ success: false });
+    }
+    res.status(200).json({
+      succes: true,
+      boards: result,
+    });
+  });
+};
