+++ routes/user/logout.js
@@ -1,0 +1,6 @@
+module.exports = (req, res) => {
+  req.user.deleteToken(req.token, (err, user) => {
+    if (err) return res.status(400).send(err);
+    res.sendStatus(200);
+  });
+};
