+++ routes/user/register.js
@@ -1,0 +1,31 @@
+const User = require("../../models/user");
+module.exports = (req, res) => {
+  // taking a user
+  const newuser = new User({
+    email: req.body.email,
+    firstname: req.body.firstname,
+    lastname: req.body.lastname,
+    password: req.body.password,
+    password2: req.body.password2,
+    image: req.file.path,
+  });
+
+  if (newuser.password != newuser.password2)
+    return res.status(400).json({ message: "password not match" });
+
+  User.findOne({ email: newuser.email }, function (err, user) {
+    if (user)
+      return res.status(400).json({ auth: false, message: "email exits" });
+
+    newuser.save((err, doc) => {
+      if (err) {
+        console.log(err);
+        return res.status(400).json({ success: false });
+      }
+      res.status(200).json({
+        succes: true,
+        user: doc,
+      });
+    });
+  });
+};
